Recover world and cloud render interfaces

+++ src/main/dll/dll_0011_screens.c
@@ -67,7 +67,7 @@ int hintTextMapFn_800ea264(void)
     return ret;
 }
 
-u8 getNextTaskHintText(void)
+u16 getNextTaskHintText(void)
 {
     u8* p = getLastSavedGameTexts();
     return p[5];
